Add classes 'linked_{ptr,array,key}<>'.
These are thread-unsafe, reference-counted pointers which keep their refcounts
by way of a circularly- and double-linked list of the participating pointers.
linked_ptr<> and linked_array<> are counterparts to scoped_{ptr,array}<>.

linked_key<> is a separate innovation, that allows non-copyable types to be
used as the keys of associative containers such as std::{map,set}<>.  There is
currently a lot of ugly copying, which I would rather get rid of.